Move _Node class to itsim/machine/__init__.py
Also created the _Socket ABC for Socket, which interacts with _Node.
