+++ itsim/machine/process_management/process.py
@@ -1,5 +1,5 @@
 from itsim.simulator import Simulator
-from itsim.machine.node import _Node
+from itsim.machine.__init__ import _Node
 from itsim.machine.process_management import _Process
 from itsim.machine.process_management.thread import Thread
 from itsim.types import Interrupt
